Reimplement beforeAllCallback as a part of test-support - Improved naming - New 'base' class for tests integrating with octopus server - Refactors AT tests to use new base class from test-support
