+++ acceptance-tests/src/test/java/com/octopus/sdk/test/ProjectsAcceptanceTest.java
@@ -26,6 +26,7 @@ import com.octopus.sdk.model.project.ProjectResource;
 import com.octopus.sdk.model.project.ProjectResourceWithLinks;
 import com.octopus.sdk.model.spaces.SpaceHome;
 import com.octopus.sdk.model.spaces.SpaceOverviewWithLinks;
+import com.octopus.testsupport.BaseOctopusServerIntegrationTest;
 
 import java.io.IOException;
 import java.net.URL;
@@ -33,7 +34,7 @@ import java.net.URL;
 import com.google.common.collect.Sets;
 import org.junit.jupiter.api.Test;
 
-public class ProjectsAcceptanceTest extends BaseAcceptanceTest {
+public class ProjectsAcceptanceTest extends BaseOctopusServerIntegrationTest {
 
   @Test
   public void returnsAProjectApiForAKnownSpace() throws IOException {
