+++ acceptance-tests/src/test/java/com/octopus/sdk/test/SpaceScopedAcceptanceTest.java
@@ -21,6 +21,7 @@ import com.octopus.sdk.http.OctopusClient;
 import com.octopus.sdk.http.RequestEndpoint;
 import com.octopus.sdk.model.spaces.SpaceHome;
 import com.octopus.sdk.model.spaces.SpaceOverviewWithLinks;
+import com.octopus.testsupport.BaseOctopusServerIntegrationTest;
 
 import java.io.IOException;
 import java.net.URL;
@@ -31,7 +32,7 @@ import org.apache.logging.log4j.Logger;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 
-public class SpaceScopedAcceptanceTest extends BaseAcceptanceTest {
+public class SpaceScopedAcceptanceTest extends BaseOctopusServerIntegrationTest {
 
   private static final Logger LOG = LogManager.getLogger();
