+++ acceptance-tests/src/test/java/com/octopus/sdk/test/SpacesHomeAcceptanceTest.java
@@ -21,6 +21,7 @@ import com.octopus.sdk.api.SpaceHomeApi;
 import com.octopus.sdk.api.SpacesOverviewApi;
 import com.octopus.sdk.http.OctopusClient;
 import com.octopus.sdk.model.spaces.SpaceOverviewWithLinks;
+import com.octopus.testsupport.BaseOctopusServerIntegrationTest;
 
 import java.io.IOException;
 import java.net.URL;
@@ -29,7 +30,7 @@ import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
 
-public class SpacesHomeAcceptanceTest extends BaseAcceptanceTest {
+public class SpacesHomeAcceptanceTest extends BaseOctopusServerIntegrationTest {
 
   @Test
   public void throwsExceptionIfRequestSpaceDoesNotExist() throws IOException {
