+++ acceptance-tests/src/test/java/com/octopus/sdk/test/SpacesOverviewAcceptanceTest.java
@@ -26,6 +26,7 @@ import com.octopus.sdk.http.HttpException;
 import com.octopus.sdk.http.OctopusClient;
 import com.octopus.sdk.model.spaces.SpaceOverviewResource;
 import com.octopus.sdk.model.spaces.SpaceOverviewWithLinks;
+import com.octopus.testsupport.BaseOctopusServerIntegrationTest;
 
 import java.io.IOException;
 import java.net.MalformedURLException;
@@ -40,7 +41,7 @@ import org.assertj.core.util.Lists;
 import org.assertj.core.util.Sets;
 import org.junit.jupiter.api.Test;
 
-public class SpacesOverviewAcceptanceTest extends BaseAcceptanceTest {
+public class SpacesOverviewAcceptanceTest extends BaseOctopusServerIntegrationTest {
 
   @Test
   public void throwsHttpExceptionIndicatingNotAuthorisedIfIncorrectApiKey()
