+++ acceptance-tests/src/test/java/com/octopus/sdk/test/UserAcceptanceTest.java
@@ -21,13 +21,14 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import com.octopus.sdk.api.UsersApi;
 import com.octopus.sdk.http.HttpException;
 import com.octopus.sdk.http.OctopusClient;
+import com.octopus.testsupport.BaseOctopusServerIntegrationTest;
 
 import java.io.IOException;
 import java.net.URL;
 
 import org.junit.jupiter.api.Test;
 
-public class UserAcceptanceTest extends BaseAcceptanceTest {
+public class UserAcceptanceTest extends BaseOctopusServerIntegrationTest {
 
   private OctopusClient client;
