+++ test-support/src/main/java/com/octopus/testsupport/OctopusDeployServerFactory.java
@@ -28,7 +28,7 @@ public class OctopusDeployServerFactory {
   private static final String USE_EXISTING_ENV_VAR_NAME = "OCTOPUS_SDK_AT_USE_EXISTING_SERVER";
 
   // Adjust these values if using a pre-running octopus server.
-  private static final boolean DEFAULT_USE_EXISTING_SERVER = true;
+  private static final boolean DEFAULT_USE_EXISTING_SERVER = false;
   protected static String serverURL = "http://localhost:8065";
   protected static String apiKey = "API-D62EQ9I4EVET1E2LJUBKEHLNBYWMO3";
