+++ test-support/src/main/java/com/octopus/testsupport/OctopusServerResourceExtension.java
@@ -1,22 +1,7 @@
-/*
- * Copyright (c) Octopus Deploy and contributors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use
- * these files except in compliance with the License. You may obtain a copy of the
- * License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed
- * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
- * CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
+package com.octopus.testsupport;
 
-package com.octopus.sdk.test;
-
-import com.octopus.testsupport.OctopusDeployServer;
-import com.octopus.testsupport.OctopusDeployServerFactory;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReentrantLock;
 
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
@@ -24,13 +9,20 @@ import org.junit.jupiter.api.extension.ParameterContext;
 import org.junit.jupiter.api.extension.ParameterResolutionException;
 import org.junit.jupiter.api.extension.ParameterResolver;
 
-public class ServerResourceExtension implements BeforeAllCallback, ParameterResolver {
+public class OctopusServerResourceExtension implements BeforeAllCallback, ParameterResolver {
+
+  private static final Lock reentrantLock = new ReentrantLock();
 
   @Override
-  public void beforeAll(final ExtensionContext context) throws Exception {
-    ExtensionContext.Store rootStore =
-        context.getRoot().getStore(ExtensionContext.Namespace.GLOBAL);
-    rootStore.put("octopus-server", new ServerResource());
+  public void beforeAll(final ExtensionContext context) {
+    reentrantLock.lock();
+    try {
+      ExtensionContext.Store rootStore =
+          context.getRoot().getStore(ExtensionContext.Namespace.GLOBAL);
+      rootStore.put("octopus-server", new ServerResource());
+    } finally {
+      reentrantLock.unlock();
+    }
   }
 
   @Override
